Added cleanUp hook. Added more extensive log message to cleanup.py. Cleanup.py does not attempt to delete empty BlobInfos anymore.

+++ src/cleanup.py
@@ -24,16 +24,18 @@ class CleanUp(webapp.RequestHandler):
     d_images.order("timestamp")
     # iterate over query to delete all the old stuff
     for l in d_images:
-      logging.info("Query time is %s. will delete one image from saved at %s" % (query_time, l.timestamp))
+      logging.info("Query time is %s. Will delete one image from saved at %s" % (query_time, l.timestamp))
+      logging.info("Blobkey is %s." % l.blob.key())
       del_blob = blobstore.BlobInfo.get(l.blob.key())
-      del_blob.delete()
+      if del_blob:
+        del_blob.delete()
       l.delete()
 
   def post(self):
     """Simple post request handler."""
     pass
 
-application = webapp.WSGIApplication([("/cleanup", CleanUp), ],
+application = webapp.WSGIApplication([("/cleanUp", CleanUp), ],
                                      debug=False)
 
 def main():
